requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
python-dotenv
sqlalchemy
psycopg2-binary
boto3
pydantic-settings
